requirements.txt
asgiref==3.11.1
django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
pyjwt==2.12.1
sqlparse==0.5.5
tzdata==2025.3
